backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy
pymysql
psycopg2-binary
python-dotenv
cryptography
pyjwt
python-multipart
fastapi-sso
httpx
